# openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coursaty API",
    "description": "Backend API for Coursaty – online course platform (مشروع التخرج)",
    "version": "1.0.0"
  },
  "paths": {
    "/api/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "operationId": "register_api_register_post",
        "requestBody": {
          "content": {
            "application/json": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": "@{application/json=}"
          },
          "422": {
            "description": "Validation Error",
            "content": "@{application/json=}"
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "OAuth2 compatible login. The 'username' field accepts the user's email.\nUsed by Swagger UI Authorize button.",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": "@{application/json=}"
          },
          "422": {
            "description": "Validation Error",
            "content": "@{application/json=}"
          }
        }
      }
    },
...
    "/api/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Me",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": "@{application/json=}"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": ""
          }
        ]
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": "@{application/json=}"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": ""
          }
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete User",
        "operationId": "delete_user_api_admin_users__user_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": ""
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; title=User Id}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": "@{application/json=}"
          },
          "422": {
            "description": "Validation Error",
            "content": "@{application/json=}"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": "",
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}